refactor: chatting-redis 브랜치에서 Yn 클래스 생성 및 경로 수정
